+++ src/Exception/InvalidSchemaException.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Phector;
+namespace Phector\Exception;
 
 /**
  * An exception to indicate a schema is invalid.
