+++ src/Exception/RecordNotFoundException.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Phector;
+namespace Phector\Exception;
 
 final class RecordNotFoundException extends \Exception
 {
